restructured auth steps and made ui changes
