+++ frontend/package.json
@@ -30,5 +30,8 @@
     "react-toastify": "^9.1.3",
     "tailwindcss": "3.3.2",
     "use-places-autocomplete": "^4.0.0"
+  },
+  "devDependencies": {
+    "encoding": "^0.1.13"
   }
 }
